+++ test/staleness-test.js
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+// @slow
 /**
  * Staleness detection unit tests.
  */